Update 2021_12_22__000_Article_VAOP_Shopping_Cart_calculator.py
va-script is a group of Actions. Action belongs to the some group
Now we can work with group of Actions
va-script is a group of actions may contain another group on the same level. Prefix 'sub' to the group is not working here. Actions and Groups of Actions are on one level.